controllers modificados y funcionando
