+++ src/routes/SessionsRouter.js
@@ -1,7 +1,6 @@
-import jwt from "jsonwebtoken";
 import passportCall from "../middlewares/passportCall.js";
 import BaseRouter from "./BaseRouter.js";
-import config from "../config/config.js";
+import sessionsController from "../controllers/sessions.controller.js";
 
 class SessionsRouter extends BaseRouter {
   init() {
@@ -9,39 +8,19 @@ class SessionsRouter extends BaseRouter {
       "/register",
       ["PUBLIC"],
       passportCall("register", { strategyType: "LOCALS" }),
-      async (req, res) => {
-        res.clearCookie("cart");
-        return res.sendSuccess("Registered");
-      }
+      sessionsController.register
     );
+
     this.post(
       "/login",
       ["NO_AUTH"],
       passportCall("login", { strategyType: "LOCALS" }),
-      async (req, res) => {
-        const tokenizedUser = {
-          name: `${req.user.firstName} ${req.user.lastName}`,
-          id: req.user._id,
-          role: req.user.role,
-          cart: req.user.cart,
-          email: req.user.email,
-        };
-        const token = jwt.sign(tokenizedUser, config.jwt.SECRET, {
-          expiresIn: "1d",
-        });
-        res.cookie(config.jwt.COOKIE, token);
-        res.clearCookie("cart");
-        return res.sendSuccess("Logged In");
-      }
+      sessionsController.login
     );
-    this.get("/logout", ["AUTH"], async (req, res) => {
-      res.clearCookie(config.jwt.COOKIE);
-      return res.sendSuccess("Logged Out");
-    });
 
-    this.get("/current", ["AUTH"], async (req, res) => {
-      return res.sendSuccessWithPayload(req.user);
-    });
+    this.get("/logout", ["AUTH"], sessionsController.logout);
+
+    this.get("/current", ["AUTH"], sessionsController.current);
   }
 }
